+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added documentation for `update-config` command.
+- Generate Ed25519 key for instance actor.
 
 ### Changed